Split RAPIDS docs extension into modules
