+++ extensions/rapids_docs/__init__.py
@@ -1,0 +1,16 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Data-driven rendering and notice support for the RAPIDS documentation portal."""
+
+from .lifecycle import _builder_inited, _source_read
+from .notices import _build_rss
+from .output import _copy_portal_files
+
+
+def setup(app):
+    app.connect("builder-inited", _builder_inited)
+    app.connect("source-read", _source_read)
+    app.connect("build-finished", _build_rss)
+    app.connect("build-finished", _copy_portal_files)
+    return {"version": "1.0", "parallel_read_safe": False, "parallel_write_safe": True}
