+++ pyproject.toml
@@ -37,7 +37,7 @@ target-version = "py312"
 select = ["B", "E", "F", "I", "UP"]
 
 [tool.ruff.lint.per-file-ignores]
-"extensions/rapids_docs.py" = ["E501"]
+"extensions/rapids_docs/*.py" = ["E501"]
 
 [tool.pytest.ini_options]
 pythonpath = ["."]
