+++ tests/test_rendering.py
@@ -5,38 +5,47 @@ from pathlib import Path
 from types import SimpleNamespace
 
 from extensions import rapids_docs
+from extensions.rapids_docs import api, lifecycle, notices, platform_support, releases
+from extensions.rapids_docs import data as portal_data
 
 ROOT = Path(__file__).resolve().parents[1]
 APP = SimpleNamespace(srcdir=str(ROOT))
 
 
-def test_data_driven_content() -> None:
-    data = rapids_docs._load_data(APP)
+def test_api_docs() -> None:
+    data = portal_data._load_data(APP)
     stable_version = data["releases"]["stable"]["version"]
     nightly_version = data["releases"]["nightly"]["version"]
     legacy_version = data["releases"]["legacy"]["version"]
 
-    api = rapids_docs._api_docs(data, "apis")
-    assert f"Stable ({stable_version})" in api
-    assert "/api/cudf/stable/" in api
-    assert "::::{grid} 1 1 1 1" in api
-    assert ":::{grid-item-card} cuDF" in api
-    assert "**Documentation:**" in api
-    assert "[GitHub]" in api
-    assert "DOCS" not in api
-    assert api.index(f"Nightly ({nightly_version})") < api.index(f"Stable ({stable_version})")
-    assert api.index(f"Stable ({stable_version})") < api.index(f"Legacy ({legacy_version})")
+    rendered = api._api_docs(data, "apis")
+    assert f"Stable ({stable_version})" in rendered
+    assert "/api/cudf/stable/" in rendered
+    assert "::::{grid} 1 1 1 1" in rendered
+    assert ":::{grid-item-card} cuDF" in rendered
+    assert "**Documentation:**" in rendered
+    assert "[GitHub]" in rendered
+    assert "DOCS" not in rendered
+    assert rendered.index(f"Nightly ({nightly_version})") < rendered.index(
+        f"Stable ({stable_version})"
+    )
+    assert rendered.index(f"Stable ({stable_version})") < rendered.index(
+        f"Legacy ({legacy_version})"
+    )
 
-    inactive = rapids_docs._api_docs(data, "inactive-projects")
+    inactive = api._api_docs(data, "inactive-projects")
     inactive_project = next(
         project
         for project in data["docs"]["inactive-projects"].values()
         if not project.get("hidden", False) and project["versions"].get("stable") == 1
     )
-    inactive_version = rapids_docs._version_label(inactive_project, "stable", data["releases"])
+    inactive_version = api._version_label(inactive_project, "stable", data["releases"])
     assert f"Stable ({inactive_version})" in inactive
 
-    platform = rapids_docs._platform_support(data)
+
+def test_platform_support() -> None:
+    data = portal_data._load_data(APP)
+    platform = platform_support._platform_support(data)
     platform_release = data["platform_support"]["releases"][0]
     assert f"RAPIDS {platform_release['version']}" in platform
     assert f"CUDA {platform_release['cuda'][0]['major']}" in platform
@@ -47,41 +56,66 @@ def test_data_driven_content() -> None:
     assert 'class="fas fa-microchip"' in platform
     assert 'class="fas fa-hammer"' in platform
 
-    schedules = rapids_docs._current_schedules(data)
+
+def test_release_schedules() -> None:
+    data = portal_data._load_data(APP)
+    nightly_version = data["releases"]["nightly"]["version"]
+    schedules = releases._current_schedules(data)
     assert f"Release v{nightly_version} Schedule" in schedules
     assert "PROPOSED" in schedules
+    previous_version = data["previous_releases"][0]["version"]
+    assert f"Release v{previous_version} Schedule" in releases._previous_schedules(data)
 
 
 def test_standard_jinja_syntax_and_raw_blocks() -> None:
     app = SimpleNamespace(srcdir=str(ROOT))
-    app.rapids_portal_data = rapids_docs._load_data(app)
-    template = rapids_docs._jinja_environment(app).from_string(
+    app.rapids_portal_data = portal_data._load_data(app)
+    template = lifecycle._jinja_environment(app).from_string(
         "{{ releases.stable.version }}\n{% raw %}${{ matrix.PY_VER }}{% endraw %}\n"
     )
 
-    rendered = template.render(rapids_docs._context(app))
+    rendered = template.render(lifecycle._context(app))
     stable_version = app.rapids_portal_data["releases"]["stable"]["version"]
 
     assert rendered == stable_version + "\n${{ matrix.PY_VER }}\n"
 
 
 def test_notice_metadata_and_indexes() -> None:
-    data = rapids_docs._load_data(APP)
+    data = portal_data._load_data(APP)
     source_notices = list((ROOT / "notices").glob("r[dgs]n[0-9][0-9][0-9][0-9].md"))
     assert len(data["notices"]) == len(source_notices)
 
-    support_notices = rapids_docs._notice_table(data, "rsn")
+    support_notices = notices._notice_table(data, "rsn")
     support_notice = next(notice for notice in data["notices"] if notice["notice_type"] == "rsn")
     assert f"RSN {support_notice['notice_id']}" in support_notices
     assert support_notice["title"] in support_notices
 
-    pinned = rapids_docs._notice_table(data, pinned=True)
+    pinned = notices._notice_table(data, pinned=True)
     pinned_notice = next(notice for notice in data["notices"] if notice.get("notice_pin"))
     assert f"{pinned_notice['notice_type'].upper()} {pinned_notice['notice_id']}" in pinned
 
 
 def test_github_alert_conversion() -> None:
     source = "> [!WARNING]\n> Do not report security vulnerabilities publicly!\n"
-    converted = rapids_docs._convert_github_alerts(source)
+    converted = lifecycle._convert_github_alerts(source)
 
     assert converted == ("```{warning}\nDo not report security vulnerabilities publicly!\n```\n")
+
+
+def test_extension_setup() -> None:
+    connections = []
+    app = SimpleNamespace(connect=lambda event, callback: connections.append((event, callback)))
+
+    metadata = rapids_docs.setup(app)
+
+    assert [event for event, _ in connections] == [
+        "builder-inited",
+        "source-read",
+        "build-finished",
+        "build-finished",
+    ]
+    assert metadata == {
+        "version": "1.0",
+        "parallel_read_safe": False,
+        "parallel_write_safe": True,
+    }
